+++ apps/frontend/src/libs/modules/map-client/libs/types/types.ts
@@ -1,6 +1,5 @@
 export { type MapMarkerOptions } from "./map-marker-options.type.js";
 export { type MapMarker } from "./map-marker.type.js";
-export { type Coordinates } from "@smartscapes/shared";
 export {
 	type ControlPosition,
 	type GeolocateControlOptions,