+++ apps/frontend/src/modules/points-of-interest/points-of-interest.ts
@@ -14,6 +14,7 @@ export {
 	type PointOfInterestPatchRequestDto,
 	type PointsOfInterestCreateRequestDto,
 	type PointsOfInterestGetAllItemResponseDto,
+	type PointsOfInterestGetAllQuery,
 	type PointsOfInterestGetByIdResponseDto,
 } from "./libs/types/types.js";
 export { pointOfInterestApi };